+++ backend/src/main/java/com/plog/backend/domain/neighbor/controller/NeighborController.java
@@ -6,6 +6,7 @@ import com.plog.backend.domain.neighbor.dto.response.NeighborCheckResponseDto;
 import com.plog.backend.domain.neighbor.dto.response.NeighborFromResponseDto;
 import com.plog.backend.domain.neighbor.dto.response.NeighborToResponseDto;
 import com.plog.backend.domain.neighbor.service.NeighborServiceImpl;
+import com.plog.backend.global.exception.NoTokenRequestException;
 import com.plog.backend.global.exception.NotValidRequestException;
 import com.plog.backend.global.model.response.BaseResponseBody;
 import io.swagger.v3.oas.annotations.Operation;
@@ -31,6 +32,8 @@ public class NeighborController {
     @Operation(summary = "이웃 추가", description = "이웃을 추가합니다.")
     @PostMapping
     public ResponseEntity<?> createNeighbor(@RequestHeader("Authorization") String token, @RequestBody NeighborAddRequestDto neighborAddRequestDto) {
+        if(token.isEmpty() || token.isBlank())
+            throw new NoTokenRequestException("Access 토큰이 필요합니다.");
         neighborService.addNeighbor(token, neighborAddRequestDto.getNeighborSearchId());
         return ResponseEntity.status(HttpStatus.OK).body(BaseResponseBody.of(200, "이웃이 추가 되었습니다."));
     }
@@ -39,6 +42,8 @@ public class NeighborController {
     @Operation(summary = "이웃 삭제", description = "이웃을 삭제합니다.")
     @DeleteMapping
     public ResponseEntity<?> deleteNeighbor(@RequestHeader("Authorization") String token, @RequestBody NeighborAddRequestDto neighborAddRequestDto) {
+        if(token.isEmpty() || token.isBlank())
+            throw new NoTokenRequestException("Access 토큰이 필요합니다.");
         neighborService.deleteNeighbor(token, neighborAddRequestDto.getNeighborSearchId());
         return ResponseEntity.status(HttpStatus.OK).body(BaseResponseBody.of(200, "이웃이 삭제 되었습니다."));
     }
@@ -47,6 +52,8 @@ public class NeighborController {
     @Operation(summary = "서로 이웃 추가", description = "이웃을 추가합니다.")
     @PostMapping("/mutual/access")
     public ResponseEntity<?> createMutualNeighbor(@RequestHeader("Authorization") String token, @RequestBody NeighborMutualAddRequestDto neighborAddRequestDto) {
+        if(token.isEmpty() || token.isBlank())
+            throw new NoTokenRequestException("Access 토큰이 필요합니다.");
         neighborService.addMutualNeighbor(token, neighborAddRequestDto.getNeighborSearchId());
         return ResponseEntity.status(HttpStatus.OK).body(BaseResponseBody.of(200, "서로 이웃이 추가 되었습니다."));
     }
@@ -55,6 +62,8 @@ public class NeighborController {
     @Operation(summary = "서로 이웃 삭제", description = "이웃을 삭제합니다.")
     @DeleteMapping("/mutual")
     public ResponseEntity<?> deleteMutualNeighbor(@RequestHeader("Authorization") String token, @RequestBody NeighborMutualAddRequestDto neighborMutualAddRequestDto) {
+        if(token.isEmpty() || token.isBlank())
+            throw new NoTokenRequestException("Access 토큰이 필요합니다.");
         if(neighborMutualAddRequestDto.getIsDelete() == null)
             throw new NotValidRequestException("삭제 후 이웃 관계를 설정이 필요합니다.");
         neighborService.deleteMutualNeighbor(token, neighborMutualAddRequestDto);
@@ -109,6 +118,8 @@ public class NeighborController {
     public ResponseEntity<NeighborCheckResponseDto> checkNeighbor(
             @RequestHeader("Authorization") String token,
             @PathVariable("searchId") String searchId) {
+        if(token.isEmpty() || token.isBlank())
+            throw new NoTokenRequestException("Access 토큰이 필요합니다.");
         return ResponseEntity.ok(neighborService.checkNeighbor(token, searchId));
     }
